refactor: Make hook to separate business logic
