+++ frontend/src/pages/product/ProductPage.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useRef } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import {
   AtomicItemImage,
@@ -8,88 +8,36 @@ import {
   ItemNav,
   ParentTemplate,
 } from '../../components';
-import { logOut, useAppDispatch, useAppSelector } from '../../hooks';
-import { selectUser, useGetProductsQuery } from '../../features';
-import { IProduct, IUser } from '../../types';
+import {
+  useAddToCart,
+  useFetchProducts,
+  useGoNextPage,
+  useProductObserver,
+  useUserActionHandler,
+} from './ProductPageHook';
 
 export type ItemColRef = HTMLAnchorElement;
 
 export default function ProductPage() {
   const navigate = useNavigate();
-  const dispatch = useAppDispatch();
-  const userInfo: IUser = useAppSelector(selectUser);
 
   const { id: productId } = useParams();
-  const { data: products } = useGetProductsQuery();
-  const columnRef = useRef<ItemColRef | null>(null);
-  const [isObserving, setIsObserving] = useState(false);
+  const [product, products, currentIndex] = useFetchProducts(productId);
 
-  const goPrevPage = () => {
-    navigate(-1);
-  };
-
-  const logOutHandler = () => {
-    logOut(dispatch);
-  };
-
-  /** Current product */
-  const product =
-    products &&
-    [...products].filter(
-      (product): product is IProduct => product?._id == productId
-    )[0];
-
-  /** Current product index  */
-  const currentIndex = product ? products?.indexOf(product) : -1;
+  const [userInfo, logOutHandler, goPrevPage] = useUserActionHandler();
 
   /** Go next product page */
-  const goNextProductPage = () => {
-    if (currentIndex + 1 !== products?.length && products) {
-      navigate(`${products[currentIndex + 1]?._id}`);
-    }
-    // Go first if reaches last index
-    else if (products) {
-      navigate(`${products[0]?._id}`);
-    }
-  };
+  const goNextProductPage = () =>
+    useGoNextPage(navigate, currentIndex, products);
 
   /** go cart page with quantity */
-  const addToCart = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
+  const addToCart = (e: React.FormEvent<HTMLFormElement>) =>
+    useAddToCart(e, navigate, product);
 
-    const qty = e.currentTarget?.productSelect.value;
+  const columnRef = useRef<ItemColRef | null>(null);
 
-    if (product?.countInStock === 0) {
-      alert('Out of Stock');
-    } else {
-      navigate(`/cart/${product?._id}?qty=${qty}`);
-    }
-  };
-
-  /** Observe product column */
-  useEffect(() => {
-    const observer = new IntersectionObserver(([entry]) => {
-      if (entry.isIntersecting && columnRef.current) {
-        setIsObserving(entry.isIntersecting);
-      }
-    });
-    if (columnRef.current) {
-      observer.observe(columnRef.current);
-    }
-
-    return () => observer.disconnect();
-  }, [goNextProductPage, isObserving]);
-
-  /** Follow the observed column */
-  useEffect(() => {
-    if (isObserving && columnRef.current) {
-      columnRef.current.scrollIntoView({
-        behavior: 'smooth',
-        block: 'center',
-        inline: 'start',
-      });
-    }
-  }, [goNextProductPage, isObserving]);
+  /** Observer Product */
+  useProductObserver(columnRef, goNextProductPage);
 
   return (
     <HeaderLayout logOut={logOutHandler} userInfo={userInfo}>
